+++ visualization/app/codeCharta/ui/fileChooser/fileChooser.e2e.ts
@@ -2,6 +2,7 @@ import { goto } from "../../../puppeteer.helper"
 import { FileChooserPageObject } from "./fileChooser.po"
 import { FilePanelPageObject } from "../filePanel/filePanel.po"
 import { DialogErrorPageObject } from "../dialog/dialog.error.po"
+import { ERROR_MESSAGES } from "../../util/fileValidator"
 
 describe("FileChooser", () => {
 	let fileChooser: FileChooserPageObject
@@ -21,6 +22,7 @@ describe("FileChooser", () => {
 
 		expect(await filePanel.getSelectedName()).toEqual("sample3.cc.json")
 	})
+
 	it("should load multiple cc.json files", async () => {
 		await fileChooser.openFiles(["./app/codeCharta/assets/sample3.cc.json", "./app/codeCharta/assets/sample2.cc.json"])
 
@@ -40,7 +42,7 @@ describe("FileChooser", () => {
 
 	it("should open an invalid file, close the dialog and open a valid file", async () => {
 		await fileChooser.openFiles(["./app/codeCharta/assets/logo.png"])
-		expect(await dialogError.getMessage()).toEqual(" file is empty or invalid")
+		expect(await dialogError.getMessage()).toEqual(` ${ERROR_MESSAGES.fileIsInvalid}`)
 
 		await dialogError.clickOk()
 
@@ -51,7 +53,7 @@ describe("FileChooser", () => {
 
 	it("should open an valid and an invalid file, close the dialog and open a valid file", async () => {
 		await fileChooser.openFiles(["./app/codeCharta/assets/logo.png", "./app/codeCharta/assets/sample3.cc.json"])
-		expect(await dialogError.getMessage()).toEqual(" file is empty or invalid")
+		expect(await dialogError.getMessage()).toEqual(` ${ERROR_MESSAGES.fileIsInvalid}`)
 
 		await dialogError.clickOk()
 
@@ -63,14 +65,20 @@ describe("FileChooser", () => {
 	it("should not load a map and show error, when loading a map with warning and a map with error", async () => {
 		await fileChooser.openFiles(["./app/codeCharta/ressources/sample1_with_api_warning.cc.json", "./app/codeCharta/assets/logo.png"])
 
-		expect(await dialogError.getMessage()).toEqual(" Minor API Version Outdated")
+		expect(await dialogError.getMessage()).toEqual(` ${ERROR_MESSAGES.minorApiVersionOutdated} Found: 1.5`)
 		await dialogError.waitUntilDialogIsClosed()
 
-		expect(await dialogError.getMessage()).toEqual(" file is empty or invalid")
+		expect(await dialogError.getMessage()).toEqual(` ${ERROR_MESSAGES.fileIsInvalid}`)
 		await dialogError.clickOk()
 
 		await fileChooser.openFiles(["./app/codeCharta/assets/sample3.cc.json"], false)
 
 		expect(await filePanel.getSelectedName()).toEqual("sample3.cc.json")
+	})
+
+	it("should be able to open a cc.json with a lower minor api version without a warning", async () => {
+		await fileChooser.openFiles(["./app/codeCharta/ressources/sample1_with_lower_minor_api.cc.json"])
+
+		expect(await filePanel.getSelectedName()).toEqual("sample1_with_lower_minor_api.cc.json")
 	})
 })